[arc] - removed analyzer diagnostics to use main diagnostics package
